crawlers/konnectors-crawler/requirements.txt
arrow
gitpython
requests
subprocess32
